add back update step for now
