+++ examples/example2.py
@@ -20,7 +20,7 @@ from diffusionjax.losses import get_loss
 from diffusionjax.solvers import EulerMaruyama, Annealed
 from diffusionjax.samplers import get_sampler
 from diffusionjax.models import CNN
-from diffusionjax.utils import get_score, retrain_nn
+from diffusionjax.utils import get_score, retrain_nn, optimizer, update_step
 from diffusionjax.sde import VP, UDLangevin
 from mlkernels import Matern52
 import numpy as np
@@ -65,30 +65,6 @@ def plot_samples_1D(samples, image_size, fname="samples 1D.png"):
     plt.plot(x, samples[:, :, 0, 0].T)
     plt.savefig(fname)
     plt.close()
-
-
-#Initialize the optimizer
-optimizer = optax.adam(1e-3)
-
-
-@partial(jit, static_argnums=[4])
-def update_step(params, rng, batch, opt_state, loss):
-    """
-    Takes the gradient of the loss function and updates the model weights (params) using it.
-    Args:
-        params: the current weights of the model
-        rng: random number generator from jax
-        batch: a batch of samples from the training data, representing samples from \mu_text{data}, shape (J, N)
-        opt_state: the internal state of the optimizer
-        loss: A loss function that can be used for score matching training.
-    Returns:
-        The value of the loss function (for metrics), the new params and the new optimizer states function (for metrics),
-        the new params and the new optimizer state.
-    """
-    val, grads = value_and_grad(loss)(params, rng, batch)
-    updates, opt_state = optimizer.update(grads, opt_state)
-    params = optax.apply_updates(params, updates)
-    return val, params, opt_state
 
 
 def main():
